pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bluetool"
version = "1.0.0"
description = "Blue Team Toolkit - Defensive Security Monitor for Home/Lab Networks"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Van Chung Nguyen", email = "ngv.chungg@gmail.com"}
]
maintainers = [
    {name = "Van Chung Nguyen", email = "ngv.chungg@gmail.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking :: Monitoring",
]
keywords = ["security", "monitoring", "network", "blue-team", "intrusion-detection"]
requires-python = ">=3.10"
dependencies = [
    "scapy>=2.5.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
geoip = ["geoip2>=4.7.0"]
performance = ["uvloop>=0.19.0; sys_platform != 'win32'"]
dev = [
    "ruff>=0.1.6",
    "black>=23.11.0",
    "pytest>=7.4.3",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/vanchungnguyxn/blue-team-toolkits"
Repository = "https://github.com/vanchungnguyxn/blue-team-toolkits"
Issues = "https://github.com/vanchungnguyxn/blue-team-toolkits/issues"

[project.scripts]
bluetool = "bluetool.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ["py310"]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
